Broke integration tests into modules
