--- tests/element_test_vectors.rs
@@ -1,869 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates.
-#![cfg(feature = "experimental-reader")]
-#![cfg(feature = "experimental-writer")]
-
-use ion_rs::{
-    BinaryWriterBuilder, Element, ElementReader, ElementWriter, Format, IonData, IonError,
-    IonResult, IonWriter, SExp, Sequence, Symbol, TextKind, TextWriterBuilder, Value,
-};
-
-use std::fs::read;
-use std::path::MAIN_SEPARATOR as PATH_SEPARATOR;
-use test_generator::test_resources;
-
-/// Concatenates two slices of string slices together.
-#[inline]
-fn concat<'a>(left: &[&'a str], right: &[&'a str]) -> Vec<&'a str> {
-    left.iter().chain(right.iter()).copied().collect()
-}
-
-/// Determines if the given file name is in the paths list.  This deals with platform
-/// path separator differences from '/' separators in the path list.
-#[inline]
-fn contains_path(paths: &[&str], file_name: &str) -> bool {
-    paths
-        .iter()
-        // TODO construct the paths in a not so hacky way
-        .map(|p| p.replace('/', &PATH_SEPARATOR.to_string()))
-        .any(|p| p == file_name)
-}
-
-// Statically defined arrays containing test file paths to skip in various contexts.
-type SkipList = &'static [&'static str];
-
-trait RoundTrip {
-    /// Encodes `elements` to a buffer in the specified Ion `format` and then reads them back into
-    /// a `Vec<Element>` using the provided reader.
-    fn roundtrip<R: ElementReader>(
-        elements: &[Element],
-        format: Format,
-        reader: R,
-    ) -> IonResult<Vec<Element>>;
-}
-
-/// Serializes all of the given [Element]s to a `Vec<u8>` according to the
-/// specified [Format].
-fn serialize(format: Format, elements: &Sequence) -> IonResult<Vec<u8>> {
-    let mut buffer = Vec::with_capacity(2048);
-    match format {
-        Format::Text(kind) => {
-            let mut writer = match kind {
-                TextKind::Compact => TextWriterBuilder::default().build(&mut buffer),
-                TextKind::Lines => TextWriterBuilder::lines().build(&mut buffer),
-                TextKind::Pretty => TextWriterBuilder::pretty().build(&mut buffer),
-                _ => unimplemented!("No text writer available for requested TextKind {:?}", kind),
-            }?;
-            writer.write_elements(elements)?;
-            writer.flush()?;
-        }
-        Format::Binary => {
-            let mut binary_writer = BinaryWriterBuilder::new().build(&mut buffer)?;
-            binary_writer.write_elements(elements)?;
-            binary_writer.flush()?;
-        }
-        _ => unimplemented!("requested format '{:?}' is not supported", format),
-    };
-    Ok(buffer)
-}
-
-trait ElementApi {
-    type ElementReader<'a>: ElementReader;
-
-    fn global_skip_list() -> SkipList;
-    fn read_one_equivs_skip_list() -> SkipList;
-    fn round_trip_skip_list() -> SkipList;
-    fn equivs_skip_list() -> SkipList;
-    fn non_equivs_skip_list() -> SkipList;
-
-    fn make_reader(data: &[u8]) -> IonResult<Self::ElementReader<'_>>;
-
-    /// Asserts the given elements can be round-tripped and equivalent, then returns the new elements.
-    fn assert_round_trip(source_elements: &Sequence, format: Format) -> IonResult<Sequence> {
-        let bytes = serialize(format, source_elements)?;
-        let mut reader = Self::make_reader(&bytes)?;
-        let new_elements = reader.read_all_elements()?;
-        assert!(
-            IonData::eq(source_elements, &new_elements),
-            "Roundtrip via {:?} failed: {}",
-            format,
-            Self::not_eq_error_message(source_elements, &new_elements)
-        );
-        Ok(new_elements)
-    }
-
-    fn not_eq_error_message(e1: &Sequence, e2: &Sequence) -> String {
-        if e1.len() != e2.len() {
-            return format!("e1 has {} elements, e2 has {} elements", e1.len(), e2.len());
-        }
-
-        for (index, (element1, element2)) in e1.iter().zip(e2.iter()).enumerate() {
-            if !IonData::eq(element1, element2) {
-                return format!(
-                    "The values at position #{index} were not IonEq.\ne1: {element1}\ne2: {element2}"
-                );
-            }
-        }
-        unreachable!("not_eq_error_message called, but all elements were ion_eq");
-    }
-
-    fn assert_three_way_round_trip(
-        file_name: &str,
-        format1: Format,
-        format2: Format,
-    ) -> IonResult<()> {
-        let source_elements = Self::read_file(file_name)?;
-        if contains_path(Self::round_trip_skip_list(), file_name) {
-            return Ok(());
-        }
-        let first_write_elements = Self::assert_round_trip(&source_elements, format1)?;
-        let second_write_elements = Self::assert_round_trip(&first_write_elements, format2)?;
-        assert!(IonData::eq(&source_elements, &second_write_elements));
-        Ok(())
-    }
-
-    fn assert_file<T, F: FnOnce() -> IonResult<T>>(
-        skip_list: &[&str],
-        file_name: &str,
-        asserter: F,
-    ) {
-        // TODO if frehberg/test-generator#7 gets implemented we could do a proper ignore
-
-        // print the paths here either way so it is easy to copy/paste to investigate failures
-        // use the --show-output argument to see it
-        if contains_path(skip_list, file_name) {
-            println!("IGNORING: {file_name}");
-        } else {
-            println!("TESTING: {file_name}");
-            asserter().unwrap();
-        }
-    }
-
-    /// Parses an element as text or binary Ion data and returns a Sequence containing all elements.
-    ///
-    /// For example, for the given group:
-    ///
-    /// ```plain
-    /// embedded_documents::(
-    ///   "a b c"
-    ///   "'a' 'b' 'c'"
-    ///   {{ 4AEA6u6OgYPbhnEDh7aBYYFigWNxCnELcQw= }}
-    /// )
-    /// ```
-    fn read_embedded_doc_as_sequence(elem: &Element) -> IonResult<Sequence> {
-        match elem.value() {
-            Value::String(text) => Element::read_all(text.text().as_bytes()),
-            Value::Blob(bytes) => Element::read_all(bytes.as_ref()),
-            _ => panic!("Expected embedded document to be an Ion String or Ion Blob"),
-        }
-    }
-
-    /// Parses a document that has top-level `list`/`sexp` values that represent a *group*.
-    /// If this top-level value is annotated with `embedded_documents`, then all values in the group
-    /// are interpreted as an embedded document, and read using [`read_embedded_doc_as_sequence`].
-    /// (The returned Sequence is then wrapped in a SExp so they can be compared as Elements).
-    /// After possibly handling an embedded document, `value_assert` function is invoked for
-    /// every pair of values in a group including the identity case (a value in a group is compared
-    /// against itself).
-    ///
-    /// Here is an example with two groups that might be used for equivalence testing:
-    ///
-    /// ```plain
-    /// (name $2 'name')
-    /// embedded_documents::(
-    ///   "a name"
-    ///   "a $2"
-    ///   "'a' 'name'"
-    /// )
-    /// ```
-    ///
-    /// This would have two groups, one with direct values that will be compared and another
-    /// with embedded Ion text that will be parsed and compared.
-    fn read_group<F>(file_name: &str, value_assert: F) -> IonResult<()>
-    where
-        // group index, value 1 name/index, value 1, value 2 name/index, value 2
-        F: Fn(usize, String, &Element, String, &Element),
-    {
-        let group_lists = Self::read_file(file_name)?;
-        for (group_index, group_container) in group_lists.iter().enumerate() {
-            // every grouping set is a list/sexp/struct
-            // look for the embedded annotation to parse/test as the underlying value
-            let is_embedded = group_container.annotations().contains("embedded_documents");
-            match group_container.value() {
-                Value::List(group) | Value::SExp(group) => {
-                    let group = if is_embedded {
-                        group
-                            .iter()
-                            .map(|it| Self::read_embedded_doc_as_sequence(it).unwrap())
-                            .map(|it| Element::from(SExp::from(it)))
-                            .collect::<Vec<_>>()
-                            .into()
-                    } else {
-                        group.to_owned()
-                    };
-
-                    for (this_index, this) in group.elements().enumerate() {
-                        for (that_index, that) in group.elements().enumerate() {
-                            value_assert(
-                                group_index,
-                                format!("{this_index}"),
-                                this,
-                                format!("{that_index}"),
-                                that,
-                            );
-                        }
-                    }
-                }
-                Value::Struct(group) => {
-                    let fields: Vec<(Symbol, Element)> = group
-                        .fields()
-                        .map(|(fname, value)| {
-                            let value = if is_embedded {
-                                let seq = Self::read_embedded_doc_as_sequence(value).unwrap();
-                                Element::from(SExp::from(seq))
-                            } else {
-                                value.to_owned()
-                            };
-                            (fname.to_owned(), value)
-                        })
-                        .collect();
-
-                    for (this_name, this) in &fields {
-                        for (that_name, that) in &fields {
-                            value_assert(
-                                group_index,
-                                format!("{this_name}"),
-                                this,
-                                format!("{that_name}"),
-                                that,
-                            );
-                        }
-                    }
-                }
-
-                _ => panic!("Expected a sequence or struct for group: {group_container:?}"),
-            };
-        }
-        Ok(())
-    }
-
-    fn read_file(file_name: &str) -> IonResult<Sequence> {
-        let data = read(file_name)?;
-        let mut reader = Self::make_reader(&data)?;
-        let result: IonResult<Sequence> = reader.read_all_elements();
-
-        // do some simple single value reading tests
-        let mut reader = Self::make_reader(&data)?;
-        let single_result: IonResult<Element> = reader.read_one_element();
-
-        match &result {
-            Ok(elems) => {
-                if elems.len() == 1 {
-                    match single_result {
-                        Ok(elem) => {
-                            // only compare if we know equality to work
-                            if !contains_path(Self::read_one_equivs_skip_list(), file_name) {
-                                assert!(IonData::eq(elems.iter().next().unwrap(), &elem))
-                            }
-                        }
-                        Err(e) => panic!("Expected element {elems:?}, got {e:?}"),
-                    }
-                } else {
-                    match single_result {
-                        Ok(elem) => {
-                            panic!("Did not expect element for duplicates: {elems:?}, {elem:?}")
-                        }
-                        Err(IonError::Decoding(_)) => (),
-                        Err(other) => {
-                            panic!("Got an error we did not expect for duplicates: {other:?}")
-                        }
-                    }
-                }
-            }
-            Err(_) => assert!(
-                single_result.is_err(),
-                "Expected error from read_one: {single_result:?}"
-            ),
-        };
-
-        result
-    }
-}
-
-/// Generates test cases for round-tripping Ion data between different formats.
-/// The arguments to the macro are a DSL that looks somewhat like Rust code. Usage:
-/// ```
-/// good_round_trip! {
-///     // Start by defining the implementation that is being tested.
-///     use NativeElementImpl;
-///     // Define some test functions—but use the actual formats rather than defining function parameters.
-///     fn my_test_1(Format::Binary, Format::Text(TextKind::Compact));
-///     fn my_test_2(Format::Binary, Format::Text(TextKind::Lines));
-/// }
-/// ```
-///
-/// ## The Generated Test Cases
-/// Reads the data in `file_name` and writes it to an in-memory buffer (`b1`) using the writer provided
-/// by `first_writer_provider`. Then reads the data from the `b1` and writes it to another buffer (`b2`)
-/// using the writer provided by `second_writer_provider`. Finally, compares the data read from `b2`
-/// to the data in `file_name` and asserts that they are equivalent, demonstrating that no data has
-/// been lost.
-macro_rules! good_round_trip {
-    (use $ElementApiImpl:ident; $(fn $test_name:ident($format1:expr, $format2:expr);)+) => {
-        mod good_round_trip {
-            use super::*; $(
-            #[test_resources("ion-tests/iontestdata_1_0/good/**/*.ion")]
-            //#[test_resources("ion-tests/iontestdata_1_1/good/**/*.ion")]
-            #[test_resources("ion-tests/iontestdata_1_0/good/**/*.10n")]
-            fn $test_name(file_name: &str) {
-                $ElementApiImpl::assert_file($ElementApiImpl::global_skip_list(), file_name, || {
-                    $ElementApiImpl::assert_three_way_round_trip(file_name, $format1, $format2)
-                });
-            })+
-        }
-    };
-}
-
-fn bad<E: ElementApi>(_element_api: E, file_name: &str) {
-    E::assert_file(E::global_skip_list(), file_name, || {
-        match E::read_file(file_name) {
-            Ok(items) => panic!("Expected error, got: {items:?}"),
-            Err(_) => Ok(()),
-        }
-    });
-}
-
-fn equivs<E: ElementApi>(_element_api: E, file_name: &str) {
-    let skip_list = concat(E::global_skip_list(), E::equivs_skip_list());
-    E::assert_file(&skip_list[..], file_name, || {
-        E::read_group(
-            file_name,
-            |group_index, this_index, this, that_index, that| {
-                assert!(
-                    IonData::eq(this,that),
-                    "in group {group_index}, index {this_index} ({this}) was not ion_eq to index {that_index} ({that})"
-                )
-            },
-        )
-    });
-}
-
-fn non_equivs<E: ElementApi>(_element_api: E, file_name: &str) {
-    let skip_list = concat(E::global_skip_list(), E::non_equivs_skip_list());
-    E::assert_file(&skip_list[..], file_name, || {
-        E::read_group(
-            file_name,
-            |group_index, this_index, this, that_index, that| {
-                if this_index == that_index {
-                    assert!(
-                        IonData::eq(this,that),
-                        "in group {group_index}, index {this_index} ({this}) was not ion_eq to index {that_index} ({that})"
-                    )
-                } else {
-                    assert!(
-                        !IonData::eq(this,that),
-                        "in group {group_index}, index {this_index} ({this}) was unexpectedly ion_eq to index {that_index} ({that})"
-                    )
-                }
-            },
-        )
-    });
-}
-
-#[cfg(test)]
-mod impl_display_for_element_tests {
-    use super::*;
-    use ion_rs::TextWriterBuilder;
-    use std::fs::read;
-
-    const TO_STRING_SKIP_LIST: &[&str] = &[
-        // These tests have shared symbol table imports in them, which the Reader does not
-        // yet support.
-        "ion-tests/iontestdata_1_0/good/subfieldVarUInt.ion",
-        "ion-tests/iontestdata_1_0/good/subfieldVarUInt15bit.ion",
-        "ion-tests/iontestdata_1_0/good/subfieldVarUInt16bit.ion",
-        "ion-tests/iontestdata_1_0/good/subfieldVarUInt32bit.ion",
-        // This test requires the reader to be able to read symbols whose ID is encoded
-        // with more than 8 bytes. Having a symbol table with more than 18 quintillion
-        // symbols is not very practical.
-        "ion-tests/iontestdata_1_0/good/typecodes/T7-large.10n",
-        "ion-tests/iontestdata_1_0/good/item1.10n",
-        "ion-tests/iontestdata_1_0/good/localSymbolTableImportZeroMaxId.ion",
-        "ion-tests/iontestdata_1_0/good/testfile35.ion",
-        // These files are encoded in utf16 and utf32; the reader currently assumes utf8.
-        "ion-tests/iontestdata_1_0/good/utf16.ion",
-        "ion-tests/iontestdata_1_0/good/utf32.ion",
-    ];
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/good/**/*.10n")]
-    fn test_to_string(file_name: &str) {
-        if contains_path(TO_STRING_SKIP_LIST, file_name) {
-            println!("IGNORING: {file_name}");
-            return;
-        }
-
-        let data = read(file_name).unwrap();
-        let result: IonResult<Sequence> = Element::read_all(data.as_slice());
-        let elements = result.unwrap_or_else(|e| {
-            panic!("Expected to be able to read Ion values for contents of file {file_name}: {e:?}")
-        });
-
-        for element in elements {
-            let mut buffer = Vec::with_capacity(2048);
-            let mut writer = TextWriterBuilder::default().build(&mut buffer).unwrap();
-            writer.write_element(&element).unwrap();
-            writer.flush().unwrap();
-            drop(writer);
-
-            let expected_string = std::str::from_utf8(buffer.as_slice()).unwrap().to_string();
-
-            assert_eq!(element.to_string(), expected_string);
-        }
-    }
-}
-
-const ELEMENT_GLOBAL_SKIP_LIST: SkipList = &[
-    // The binary reader does not check whether nested values are longer than their
-    // parent container.
-    "ion-tests/iontestdata_1_0/bad/listWithValueLargerThanSize.10n",
-    // ROUND TRIP
-    // These tests have shared symbol table imports in them, which the Reader does not
-    // yet support.
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt15bit.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt16bit.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt32bit.ion",
-    // This test requires the reader to be able to read symbols whose ID is encoded
-    // with more than 8 bytes. Having a symbol table with more than 18 quintillion
-    // symbols is not very practical.
-    "ion-tests/iontestdata_1_0/good/typecodes/T7-large.10n",
-    // ---
-    // Requires importing shared symbol tables
-    "ion-tests/iontestdata_1_0/good/item1.10n",
-    "ion-tests/iontestdata_1_0/good/localSymbolTableImportZeroMaxId.ion",
-    // Requires importing shared symbol tables
-    "ion-tests/iontestdata_1_0/good/testfile35.ion",
-    // These files are encoded in utf16 and utf32; the reader currently assumes utf8.
-    "ion-tests/iontestdata_1_0/good/utf16.ion",
-    "ion-tests/iontestdata_1_0/good/utf32.ion",
-    // NON-EQUIVS
-    "ion-tests/iontestdata_1_0/good/non-equivs/localSymbolTableWithAnnotations.ion",
-    "ion-tests/iontestdata_1_0/good/non-equivs/symbolTablesUnknownText.ion",
-];
-
-const ELEMENT_ROUND_TRIP_SKIP_LIST: SkipList = &[
-    "ion-tests/iontestdata_1_0/good/item1.10n",
-    "ion-tests/iontestdata_1_0/good/localSymbolTableImportZeroMaxId.ion",
-    "ion-tests/iontestdata_1_0/good/notVersionMarkers.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldInt.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldUInt.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarInt.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt15bit.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt16bit.ion",
-    "ion-tests/iontestdata_1_0/good/subfieldVarUInt32bit.ion",
-    "ion-tests/iontestdata_1_0/good/utf16.ion",
-    "ion-tests/iontestdata_1_0/good/utf32.ion",
-];
-
-const ELEMENT_EQUIVS_SKIP_LIST: SkipList = &[
-    "ion-tests/iontestdata_1_0/good/equivs/localSymbolTableAppend.ion",
-    "ion-tests/iontestdata_1_0/good/equivs/localSymbolTableNullSlots.ion",
-    "ion-tests/iontestdata_1_0/good/equivs/nonIVMNoOps.ion",
-];
-
-#[cfg(test)]
-mod native_element_tests {
-    use super::*;
-    use ion_rs::{Reader, ReaderBuilder};
-
-    struct NativeElementApi;
-
-    impl ElementApi for NativeElementApi {
-        type ElementReader<'a> = Reader<'a>;
-
-        fn make_reader(data: &[u8]) -> IonResult<Reader<'_>> {
-            ReaderBuilder::default().build(data)
-        }
-
-        fn global_skip_list() -> SkipList {
-            ELEMENT_GLOBAL_SKIP_LIST
-        }
-
-        fn read_one_equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn round_trip_skip_list() -> SkipList {
-            ELEMENT_ROUND_TRIP_SKIP_LIST
-        }
-
-        fn equivs_skip_list() -> SkipList {
-            ELEMENT_EQUIVS_SKIP_LIST
-        }
-
-        fn non_equivs_skip_list() -> SkipList {
-            &[]
-        }
-    }
-
-    good_round_trip! {
-        use NativeElementApi;
-        fn binary_compact(Format::Binary, Format::Text(TextKind::Compact));
-        fn binary_lines(Format::Binary, Format::Text(TextKind::Lines));
-        fn binary_pretty(Format::Binary, Format::Text(TextKind::Pretty));
-        fn compact_binary(Format::Text(TextKind::Compact), Format::Binary);
-        fn compact_lines(Format::Text(TextKind::Compact), Format::Text(TextKind::Lines));
-        fn compact_pretty(Format::Text(TextKind::Compact), Format::Text(TextKind::Pretty));
-        fn lines_binary(Format::Text(TextKind::Lines), Format::Binary);
-        fn lines_compact(Format::Text(TextKind::Lines), Format::Text(TextKind::Compact));
-        fn lines_pretty(Format::Text(TextKind::Lines), Format::Text(TextKind::Pretty));
-        fn pretty_binary(Format::Text(TextKind::Pretty), Format::Binary);
-        fn pretty_compact(Format::Text(TextKind::Pretty), Format::Text(TextKind::Compact));
-        fn pretty_lines(Format::Text(TextKind::Pretty), Format::Text(TextKind::Lines));
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.10n")]
-    fn native_bad(file_name: &str) {
-        bad(NativeElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.10n")]
-    fn native_equivs(file_name: &str) {
-        equivs(NativeElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.ion")]
-    // no binary files exist and the macro doesn't like empty globs...
-    // see frehberg/test-generator#12
-    //#[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.10n")]
-    fn native_non_equivs(file_name: &str) {
-        non_equivs(NativeElementApi, file_name)
-    }
-}
-
-#[cfg(test)]
-mod non_blocking_native_element_tests {
-    use super::*;
-    use ion_rs::{RawBinaryReader, RawTextReader, Reader};
-
-    struct NonBlockingNativeElementApi;
-
-    impl ElementApi for NonBlockingNativeElementApi {
-        type ElementReader<'a> = Reader<'a>;
-
-        fn make_reader(data: &[u8]) -> IonResult<Self::ElementReader<'_>> {
-            // These imports are visible as a temporary workaround.
-            // See: https://github.com/amazon-ion/ion-rust/issues/484
-            use ion_rs::integration_testing::{new_reader, IVM};
-            // If the data is binary, create a non-blocking binary reader.
-            if data.starts_with(&IVM) {
-                let raw_reader = RawBinaryReader::new(data);
-                Ok(new_reader(raw_reader))
-            } else {
-                // Otherwise, create a non-blocking text reader
-                let raw_reader = RawTextReader::new(data);
-                Ok(new_reader(raw_reader))
-            }
-        }
-
-        fn global_skip_list() -> SkipList {
-            ELEMENT_GLOBAL_SKIP_LIST
-        }
-
-        fn read_one_equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn round_trip_skip_list() -> SkipList {
-            ELEMENT_ROUND_TRIP_SKIP_LIST
-        }
-
-        fn equivs_skip_list() -> SkipList {
-            ELEMENT_EQUIVS_SKIP_LIST
-        }
-
-        fn non_equivs_skip_list() -> SkipList {
-            &[]
-        }
-    }
-
-    good_round_trip! {
-        use NonBlockingNativeElementApi;
-        fn binary_compact(Format::Binary, Format::Text(TextKind::Compact));
-        fn binary_lines(Format::Binary, Format::Text(TextKind::Lines));
-        fn binary_pretty(Format::Binary, Format::Text(TextKind::Pretty));
-        fn compact_binary(Format::Text(TextKind::Compact), Format::Binary);
-        fn compact_lines(Format::Text(TextKind::Compact), Format::Text(TextKind::Lines));
-        fn compact_pretty(Format::Text(TextKind::Compact), Format::Text(TextKind::Pretty));
-        fn lines_binary(Format::Text(TextKind::Lines), Format::Binary);
-        fn lines_compact(Format::Text(TextKind::Lines), Format::Text(TextKind::Compact));
-        fn lines_pretty(Format::Text(TextKind::Lines), Format::Text(TextKind::Pretty));
-        fn pretty_binary(Format::Text(TextKind::Pretty), Format::Binary);
-        fn pretty_compact(Format::Text(TextKind::Pretty), Format::Text(TextKind::Compact));
-        fn pretty_lines(Format::Text(TextKind::Pretty), Format::Text(TextKind::Lines));
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.10n")]
-    fn native_bad(file_name: &str) {
-        bad(NonBlockingNativeElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.10n")]
-    fn native_equivs(file_name: &str) {
-        equivs(NonBlockingNativeElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.ion")]
-    // no binary files exist and the macro doesn't like empty globs...
-    // see frehberg/test-generator#12
-    //#[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.10n")]
-    fn native_non_equivs(file_name: &str) {
-        non_equivs(NonBlockingNativeElementApi, file_name)
-    }
-}
-
-#[cfg(all(test, feature = "experimental-streaming"))]
-mod token_native_element_tests {
-    use super::*;
-    use ion_rs::tokens::{ReaderTokenStream, TokenStreamReader};
-    use ion_rs::{Reader, ReaderBuilder};
-
-    struct TokenNativeElementApi;
-
-    impl ElementApi for TokenNativeElementApi {
-        type ElementReader<'a> = TokenStreamReader<'a, ReaderTokenStream<'a, Reader<'a>>>;
-
-        fn global_skip_list() -> SkipList {
-            ELEMENT_GLOBAL_SKIP_LIST
-        }
-
-        fn read_one_equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn round_trip_skip_list() -> SkipList {
-            ELEMENT_ROUND_TRIP_SKIP_LIST
-        }
-
-        fn equivs_skip_list() -> SkipList {
-            ELEMENT_EQUIVS_SKIP_LIST
-        }
-
-        fn non_equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn make_reader<'a>(data: &'a [u8]) -> IonResult<Self::ElementReader<'a>> {
-            let reader = ReaderBuilder::default().build(data)?;
-            let tokens: ReaderTokenStream<'a, _> = reader.into();
-            let token_reader: TokenStreamReader<'a, _> = tokens.into();
-            Ok(token_reader)
-        }
-    }
-
-    good_round_trip! {
-        use TokenNativeElementApi;
-        fn binary_compact(Format::Binary, Format::Text(TextKind::Compact));
-        fn binary_lines(Format::Binary, Format::Text(TextKind::Lines));
-        fn binary_pretty(Format::Binary, Format::Text(TextKind::Pretty));
-        fn compact_binary(Format::Text(TextKind::Compact), Format::Binary);
-        fn compact_lines(Format::Text(TextKind::Compact), Format::Text(TextKind::Lines));
-        fn compact_pretty(Format::Text(TextKind::Compact), Format::Text(TextKind::Pretty));
-        fn lines_binary(Format::Text(TextKind::Lines), Format::Binary);
-        fn lines_compact(Format::Text(TextKind::Lines), Format::Text(TextKind::Compact));
-        fn lines_pretty(Format::Text(TextKind::Lines), Format::Text(TextKind::Pretty));
-        fn pretty_binary(Format::Text(TextKind::Pretty), Format::Binary);
-        fn pretty_compact(Format::Text(TextKind::Pretty), Format::Text(TextKind::Compact));
-        fn pretty_lines(Format::Text(TextKind::Pretty), Format::Text(TextKind::Lines));
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.10n")]
-    fn native_bad(file_name: &str) {
-        bad(TokenNativeElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.10n")]
-    fn native_equivs(file_name: &str) {
-        equivs(TokenNativeElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.ion")]
-    // no binary files exist and the macro doesn't like empty globs...
-    // see frehberg/test-generator#12
-    //#[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.10n")]
-    fn native_non_equivs(file_name: &str) {
-        non_equivs(TokenNativeElementApi, file_name)
-    }
-}
-
-#[cfg(test)]
-mod lazy_element_tests {
-    use super::*;
-    use ion_rs::lazy::reader::LazyReader;
-
-    struct LazyReaderElementApi;
-
-    impl ElementApi for LazyReaderElementApi {
-        type ElementReader<'a> = LazyReader<'a>;
-
-        fn make_reader(data: &[u8]) -> IonResult<Self::ElementReader<'_>> {
-            Ok(LazyReader::new(data))
-        }
-
-        fn global_skip_list() -> SkipList {
-            ELEMENT_GLOBAL_SKIP_LIST
-        }
-
-        fn read_one_equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn round_trip_skip_list() -> SkipList {
-            ELEMENT_ROUND_TRIP_SKIP_LIST
-        }
-
-        fn equivs_skip_list() -> SkipList {
-            ELEMENT_EQUIVS_SKIP_LIST
-        }
-
-        fn non_equivs_skip_list() -> SkipList {
-            &[]
-        }
-    }
-
-    good_round_trip! {
-        use LazyReaderElementApi;
-        fn binary_compact(Format::Binary, Format::Text(TextKind::Compact));
-        fn binary_lines(Format::Binary, Format::Text(TextKind::Lines));
-        fn binary_pretty(Format::Binary, Format::Text(TextKind::Pretty));
-        fn compact_binary(Format::Text(TextKind::Compact), Format::Binary);
-        fn compact_lines(Format::Text(TextKind::Compact), Format::Text(TextKind::Lines));
-        fn compact_pretty(Format::Text(TextKind::Compact), Format::Text(TextKind::Pretty));
-        fn lines_binary(Format::Text(TextKind::Lines), Format::Binary);
-        fn lines_compact(Format::Text(TextKind::Lines), Format::Text(TextKind::Compact));
-        fn lines_pretty(Format::Text(TextKind::Lines), Format::Text(TextKind::Pretty));
-        fn pretty_binary(Format::Text(TextKind::Pretty), Format::Binary);
-        fn pretty_compact(Format::Text(TextKind::Pretty), Format::Text(TextKind::Compact));
-        fn pretty_lines(Format::Text(TextKind::Pretty), Format::Text(TextKind::Lines));
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/bad/**/*.10n")]
-    fn lazy_bad(file_name: &str) {
-        bad(LazyReaderElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.ion")]
-    #[test_resources("ion-tests/iontestdata_1_0/good/equivs/**/*.10n")]
-    fn lazy_equivs(file_name: &str) {
-        equivs(LazyReaderElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.ion")]
-    // no binary files exist and the macro doesn't like empty globs...
-    // see frehberg/test-generator#12
-    //#[test_resources("ion-tests/iontestdata_1_0/good/non-equivs/**/*.10n")]
-    fn lazy_non_equivs(file_name: &str) {
-        non_equivs(LazyReaderElementApi, file_name)
-    }
-}
-
-/// TODO: When the Ion 1.1 binary reader is complete, update this module to include binary tests
-#[cfg(test)]
-mod ion_1_1 {
-    use super::*;
-    use ion_rs::lazy::reader::LazyTextReader_1_1;
-
-    struct LazyReaderElementApi;
-
-    impl ElementApi for LazyReaderElementApi {
-        type ElementReader<'a> = LazyTextReader_1_1<'a>;
-
-        fn make_reader(data: &[u8]) -> IonResult<Self::ElementReader<'_>> {
-            Ok(LazyTextReader_1_1::new(data).unwrap())
-        }
-
-        fn global_skip_list() -> SkipList {
-            &[
-                // TODO: Revisit these once the Ion 1.1 System Symbol Table is defined
-                "ion-tests/iontestdata_1_1/good/equivs/localSymbolTableAppend.ion",
-                "ion-tests/iontestdata_1_1/good/equivs/localSymbolTableNullSlots.ion",
-                "ion-tests/iontestdata_1_1/good/equivs/localSymbolTableWithAnnotations.ion",
-                "ion-tests/iontestdata_1_1/good/equivs/localSymbolTables.ion",
-                "ion-tests/iontestdata_1_1/good/equivs/nonIVMNoOps.ion",
-                "ion-tests/iontestdata_1_1/good/non-equivs/annotatedIvms.ion",
-                "ion-tests/iontestdata_1_1/good/non-equivs/localSymbolTableWithAnnotations.ion",
-                "ion-tests/iontestdata_1_1/good/non-equivs/symbolTables.ion",
-                "ion-tests/iontestdata_1_1/good/non-equivs/symbolTablesUnknownText.ion",
-                // TODO: Remove from skiplist when shared symbol tables are supported
-                "ion-tests/iontestdata_1_1/good/localSymbolTableImportZeroMaxId.ion",
-                "ion-tests/iontestdata_1_1/good/testfile35.ion",
-                // TODO: https://github.com/amazon-ion/ion-rust/issues/653
-                "ion-tests/iontestdata_1_1/good/equivs/macros/make_string.ion",
-                "ion-tests/iontestdata_1_1/good/equivs/macros/values.ion",
-                "ion-tests/iontestdata_1_1/good/equivs/macros/void.ion",
-                "ion-tests/iontestdata_1_1/good/macros/void_invoked_deeply_nested.ion",
-                "ion-tests/iontestdata_1_1/good/macros/void_invoked_in_struct.ion",
-                "ion-tests/iontestdata_1_1/good/macros/void_invoked_in_struct_field.ion",
-            ]
-        }
-
-        fn read_one_equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn round_trip_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn equivs_skip_list() -> SkipList {
-            &[]
-        }
-
-        fn non_equivs_skip_list() -> SkipList {
-            &[]
-        }
-    }
-
-    // TODO: Update to write as Ion 1.1
-    mod good_round_trip {
-        use super::*;
-        use ion_rs::Format::Text;
-        #[test_resources("ion-tests/iontestdata_1_1/good/**/*.ion")]
-        fn round_trip(file_name: &str) {
-            LazyReaderElementApi::assert_file(
-                LazyReaderElementApi::global_skip_list(),
-                file_name,
-                || {
-                    LazyReaderElementApi::assert_three_way_round_trip(
-                        file_name,
-                        Text(TextKind::Lines),
-                        Text(TextKind::Lines),
-                    )
-                },
-            );
-        }
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_1/bad/**/*.ion")]
-    fn lazy_bad(file_name: &str) {
-        bad(LazyReaderElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_1/good/equivs/**/*.ion")]
-    fn lazy_equivs(file_name: &str) {
-        equivs(LazyReaderElementApi, file_name)
-    }
-
-    #[test_resources("ion-tests/iontestdata_1_1/good/non-equivs/**/*.ion")]
-    fn lazy_non_equivs(file_name: &str) {
-        non_equivs(LazyReaderElementApi, file_name)
-    }
-}
